# docs/generated/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/remote-assist/annotations/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "远程协助"
                ],
                "summary": "删除远程协助标注",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标注 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/remote-assist/sessions": {
            "get": {
                "description": "按会话（可选）倒序列出协助记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "远程协助"
                ],
                "summary": "获取远程协助列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "按会话过滤",
                        "name": "conversation_session_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "上限（默认 100，最大 200）",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "在指定会话上创建 active 状态的协助记录；信令与媒体面走既有 WS/RTC 通道",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "远程协助"
                ],
                "summary": "发起远程协助会话",
                "parameters": [
                    {
                        "description": "协助内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.StartAssistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.RemoteAssistSession"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/remote-assist/sessions/{id}": {
            "get": {
                "description": "按协助会话 ID 返回详情（含录制元数据）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "远程协助"
                ],
                "summary": "获取单次远程协助",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "协助会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RemoteAssistSession"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/remote-assist/sessions/{id}/annotations": {
            "get": {
                "description": "按时间戳升序返回协助会话的全部标注",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "远程协助"
                ],
                "summary": "获取远程协助标注",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "协助会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "在协助会话上追加一条 Canvas 标注（坐标按归一化比例存储）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "远程协助"
                ],
                "summary": "追加远程协助标注",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "协助会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "标注内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddAnnotationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.RemoteAssistAnnotation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/remote-assist/sessions/{id}/end": {
            "post": {
                "description": "将协助置为 ended/failed 并落录制元数据（文件本体经 /api/v1/upload 上传）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "远程协助"
                ],
                "summary": "结束远程协助会话",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "协助会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "结束内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.EndAssistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RemoteAssistSession"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.AddAnnotationRequest": {
            "type": "object"
        },
...
        "handlers.EndAssistRequest": {
            "type": "object",
            "properties": {
                "outcome": {
                    "description": "ended|failed",
                    "type": "string",
                    "example": "ended"
                },
                "recording_duration_ms": {
                    "type": "integer"
                },
                "recording_key": {
                    "type": "string"
                },
                "recording_mime": {
                    "type": "string"
                },
                "recording_size": {
                    "type": "integer"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.StartAssistRequest": {
            "type": "object",
            "required": [
                "conversation_session_id"
            ],
            "properties": {
                "agent_user_id": {
                    "type": "integer",
                    "example": 7
                },
                "conversation_session_id": {
                    "type": "string",
                    "example": "sess-123"
                },
                "tenant_id": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
        "handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.RemoteAssistAnnotation": {
            "type": "object",
            "properties": {
                "assist_session_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "payload": {
                    "type": "string"
                },
                "shape": {
                    "description": "rect|freehand|arrow",
                    "type": "string"
                },
                "timestamp_ms": {
                    "type": "integer"
                }
            }
        },
        "models.RemoteAssistSession": {
            "type": "object",
            "properties": {
                "agent_user_id": {
                    "type": "integer"
                },
                "conversation_session_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "ended_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "recording_duration_ms": {
                    "type": "integer"
                },
                "recording_key": {
                    "description": "录制元数据：文件经既有 /api/v1/upload 上传，这里只落 key 与展示信息",
                    "type": "string"
                },
                "recording_mime": {
                    "type": "string"
                },
                "recording_size": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "description": "active|ended|failed",
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
    }
}